add inContext to other context-dependent edits
